move resnet into nlc repo and rename dist to codec
