+++ craystack/core_test.py
@@ -3,9 +3,7 @@ import numpy.random as rng
 
 import craystack as cs
 import craystack.vectorans as vrans
-import craystack.distributions as cs_dist
-import craystack.bb_ans as bb
-import craystack.ar as ar
+import craystack.codecs as codecs
 
 
 def check_codec(head_shape, codec, data):
@@ -21,7 +19,7 @@ def test_uniform():
     precision = 4
     shape = (2, 3, 5)
     data = rng.randint(precision, size=shape, dtype="uint64")
-    check_codec(shape, cs_dist.Uniform(precision), data)
+    check_codec(shape, codecs.Uniform(precision), data)
 
 
 def test_benford():
@@ -29,7 +27,7 @@ def test_benford():
     data_len = rng.randint(31, 63, shape, dtype="uint64")
     data = ((1 << data_len) | (rng.randint(1 << 63, size=shape, dtype="uint64")
                                & ((1 << data_len) - 1)))
-    check_codec(shape, cs_dist.Benford64, data)
+    check_codec(shape, codecs.Benford64, data)
 
 
 def test_benford_inverse():
@@ -37,7 +35,7 @@ def test_benford_inverse():
     data_len = rng.randint(31, 63, shape, dtype="uint64")
     data = ((1 << data_len) | (rng.randint(1 << 63, size=shape, dtype="uint64")
                                & ((1 << data_len) - 1)))
-    check_codec(shape, cs_dist.Benford64, data)
+    check_codec(shape, codecs.Benford64, data)
 
 
 def test_repeat():
@@ -45,7 +43,7 @@ def test_repeat():
     n_data = 7
     shape = (2, 3, 5)
     data = rng.randint(precision, size=(n_data,) + shape, dtype="uint64")
-    check_codec(shape, cs.repeat(cs_dist.Uniform(precision), n_data), data)
+    check_codec(shape, cs.repeat(codecs.Uniform(precision), n_data), data)
 
 
 def test_substack():
@@ -56,7 +54,7 @@ def test_substack():
     message = head, tail
     data = rng.randint(1 << prec, size=(n_data, 2, 4), dtype='uint64')
     view_fun = lambda h: h[0]
-    append, pop = cs.substack(cs.repeat(cs_dist.Uniform(prec), n_data), view_fun)
+    append, pop = cs.substack(cs.repeat(codecs.Uniform(prec), n_data), view_fun)
     message_ = append(message, data)
     np.testing.assert_array_equal(message_[0][1], message[0][1])
     message_, data_ = pop(message_)
@@ -66,7 +64,7 @@ def test_substack():
 def test_parallel():
     precs = [1, 2, 4, 8, 16]
     szs = [2, 3, 4, 5, 6]
-    codecs = [cs_dist.Uniform(p) for p in precs]
+    u_codecs = [codecs.Uniform(p) for p in precs]
     view_fun = lambda slc: lambda head: head[slc]
     view_funs = []
     start = 0
@@ -75,7 +73,7 @@ def test_parallel():
         start += s
     data = [rng.randint(1 << p, size=size, dtype='uint64')
             for p, size in zip(precs, szs)]
-    check_codec(sum(szs), cs.parallel(codecs, view_funs), data)
+    check_codec(sum(szs), cs.parallel(u_codecs, view_funs), data)
 
 
 def test_bernoulli():
@@ -83,7 +81,7 @@ def test_bernoulli():
     shape = (2, 3, 5)
     p = rng.random(shape)
     data = np.uint64(rng.random(shape) < p)
-    check_codec(shape, cs_dist.Bernoulli(p, precision), data)
+    check_codec(shape, codecs.Bernoulli(p, precision), data)
 
 
 def test_categorical():
@@ -93,7 +91,7 @@ def test_categorical():
     ps = ps / np.sum(ps, axis=-1, keepdims=True)
     data = np.reshape([rng.choice(4, p=p) for p in ps], shape)
     ps = np.reshape(ps, shape + (4,))
-    check_codec(shape, cs_dist.Categorical(ps, precision), data)
+    check_codec(shape, codecs.Categorical(ps, precision), data)
 
 
 def test_logistic():
@@ -103,7 +101,7 @@ def test_logistic():
     log_scale = rng.randn() - 4
     # type is important!
     data = np.array([rng.choice(256) for _ in range(batch_size)]).astype('uint64')
-    check_codec((batch_size,), cs_dist.Logistic(means, log_scale, precision), data)
+    check_codec((batch_size,), codecs.Logistic(means, log_scale, precision), data)
 
 
 def test_logistic_mixture():
@@ -117,7 +115,7 @@ def test_logistic_mixture():
     theta = np.concatenate((means, log_scales, logit_probs), axis=-1)
     # type is important!
     data = np.array([rng.choice(256) for _ in range(batch_size)]).astype('uint64')
-    check_codec((shape[0],), cs_dist.LogisticMixture(theta, precision), data)
+    check_codec((shape[0],), codecs.LogisticMixture(theta, precision), data)
 
 
 def test_autoregressive():
@@ -130,9 +128,9 @@ def test_autoregressive():
     fixed_probs = rng.random((batch_size, data_size, choices))
     fixed_probs = fixed_probs / np.sum(fixed_probs, axis=-1, keepdims=True)
     elem_idxs = [(slice(None), i) for i in range(10)]  # slice for the batch dimension
-    elem_codec = lambda p, idx: cs_dist.Categorical(p, precision)
+    elem_codec = lambda p, idx: codecs.Categorical(p, precision)
     check_codec((batch_size,),
-                ar.AutoRegressive(lambda *x: fixed_probs,
+                codecs.AutoRegressive(lambda *x: fixed_probs,
                                   (batch_size, data_size,),
                                   fixed_probs.shape,
                                   elem_idxs,
@@ -155,7 +153,7 @@ def test_gaussian_db():
     data = np.array([rng.choice(1 << bin_precision) for _ in range(batch_size)])
 
     check_codec((batch_size,),
-                bb.DiagGaussianLatent(means, stdds, bin_means, bin_stdds,
+                codecs.DiagGaussianLatent(means, stdds, bin_means, bin_stdds,
                                       coding_precision, bin_precision),
                 data)
 
@@ -168,13 +166,13 @@ def test_flatten_unflatten_benford():
     freqs = np.ones(shape, dtype="uint64")
     for b in some_bits:
         state = vrans.append(state, b, freqs, p)
-    flat = cs_dist.flatten_benford(state)
+    flat = codecs.flatten_benford(state)
     flat_ = vrans.flatten(state)
     print('Normal flat len: {}'.format(len(flat_) * 32))
     print('Benford flat len: {}'.format(len(flat) * 32))
     assert flat.dtype is np.dtype("uint32")
-    state_ = cs_dist.unflatten_benford(flat, shape)
-    flat_ = cs_dist.flatten_benford(state_)
+    state_ = codecs.unflatten_benford(flat, shape)
+    flat_ = codecs.flatten_benford(state_)
     assert np.all(flat == flat_)
     assert np.all(state[0] == state_[0])
     assert state[1] == state_[1]
